feat: rewrite and rename cjsToEsm.cjs
